+++ plugins/kserve-v2-protocol/src/main/java/ai/djl/serving/kserve/KServeDescribeModelResponse.java
@@ -10,7 +10,7 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package ai.djl.serving.http;
+package ai.djl.serving.kserve;
 
 import ai.djl.ndarray.types.DataType;
 import ai.djl.ndarray.types.Shape;
