+++ plugins/kserve-v2-protocol/src/main/resources/META-INF/plugin.definition
@@ -1,0 +1,2 @@
+name=kserve
+export=ai.djl.serving.kserve.KServeRequestHandler
