+++ settings.gradle
@@ -6,5 +6,6 @@ include ':engines:java'
 include ':plugins:management-console'
 include ':plugins:plugin-management-plugin'
 include ':plugins:static-file-plugin'
+include 'plugins:kserve-v2-protocol'
 include ':serving'
 include ':wlm'
